Made NavBar component and added basic layout to all pages
