+++ app/(pages)/achievements/page.tsx
@@ -1,4 +1,9 @@
-const AchievementsPage = () => {
-  return <div>Achievements page</div>;
+import { Stack } from "@mui/material";
+
+export default function Achievements() {
+  return (
+    <Stack style={{ backgroundColor: "#F8EBDE", flexGrow: 1 }}>
+      Hello From Achievements
+    </Stack>
+  );
 };
-export default AchievementsPage;
