+++ app/(pages)/chat/page.tsx
@@ -1,4 +1,9 @@
-const ChatPage = () => {
-  return <div>Chat page</div>;
-};
-export default ChatPage;
+import { Stack } from "@mui/material";
+
+export default function Chat() {
+  return (
+    <Stack style={{ backgroundColor: "#F8EBDE", flexGrow: 1 }}>
+      Hello From Chat
+    </Stack>
+  );
+}
