+++ app/(pages)/dashboard/page.tsx
@@ -1,4 +1,9 @@
-const Dashboard = () => {
-    return <div>Dashboard</div>;
-  };
-  export default Dashboard;
+import { Stack } from "@mui/material";
+
+export default function Dashboard() {
+  return (
+    <Stack style={{ backgroundColor: "#F8EBDE", flexGrow: 1 }}>
+      Hello From Dashboard
+    </Stack>
+  );
+}
