+++ app/(pages)/events/page.tsx
@@ -1,4 +1,9 @@
-const EventsPage = () => {
-  return <div>Events page</div>;
-};
-export default EventsPage;
+import { Stack } from "@mui/material";
+
+export default function Events() {
+  return (
+    <Stack style={{ backgroundColor: "#F8EBDE", flexGrow: 1 }}>
+      Hello From Events
+    </Stack>
+  );
+}
