+++ app/(pages)/groups/page.tsx
@@ -1,4 +1,9 @@
-const GroupsPage = () => {
-    return <div>groups page</div>;
-  };
-  export default GroupsPage;
+import { Stack } from "@mui/material";
+
+export default function Groups() {
+  return (
+    <Stack style={{ backgroundColor: "#F8EBDE", flexGrow: 1 }}>
+      Hello From Groups
+    </Stack>
+  );
+}
