+++ app/(pages)/trips/page.tsx
@@ -1,4 +1,9 @@
-const TripsPage = () => {
-  return <div>Trips page</div>;
-};
-export default TripsPage;
+import { Stack } from "@mui/material";
+
+export default function Trips() {
+  return (
+    <Stack style={{ backgroundColor: "#F8EBDE", flexGrow: 1 }}>
+      Hello From Trips
+    </Stack>
+  );
+}
